Move components back into src
